requirements.txt
customtkinter>=5.2.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
onnxruntime>=1.15.0
transformers>=4.30.0
tkinter
